pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Paulo Roberto Cruz <paulo.cruz9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
drf-extra-fields = "^3.1.1"
Pillow = "^8.2.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
